+++ dwds/lib/src/services/chrome_proxy_service.dart
@@ -101,7 +101,6 @@ class ChromeProxyService implements VmServiceInterface {
       remoteDebugger,
       _streamNotify,
       appInspectorProvider,
-      _assetReader,
       _modules,
       _locations,
       uri,