+++ dwds/test/debugger_test.dart
@@ -38,7 +38,6 @@ void main() async {
       null,
       () => inspector,
       null,
-      null,
       locations,
       root,
     );